+++ example/test-react-dev/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@flagship.io/react-sdk": "link:../../",
+    "@flagship.io/react-sdk": "^2.2.0-alpha.2",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
