+++ example/test-react-next/package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@flagship.io/js-sdk": "^2.3.1-alpha.3",
     "@flagship.io/react-sdk": "link:../../",
     "next": "11.1.0",
     "react": "17.0.2",
