+++ example/test-react-next/pages/index.tsx
@@ -1,103 +1,31 @@
 import type { NextPage } from 'next'
-import Head from 'next/head'
-import Image from 'next/image'
-import { useFlagship } from '@flagship.io/react-sdk'
-import styles from '../styles/Home.module.css'
-import { campaigns } from '../campaigns'
+import Flagship from '@flagship.io/js-sdk'
+import { ENV_ID, API_KEY } from '../config'
+import Home from '../components/Home'
 
-let age = 1
-const Home: NextPage = () => {
-
-  const fs = useFlagship({modifications:{requested:[{key:"btnColor", defaultValue:"white"}]}})
-  
-  
-const click =()=>{
-  age = age===1?1:2
-  fs.updateContext({age})
-}
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p>{fs.modifications}</p>
-        <button style={{width:100, height:50}} value={"click me"} onClick={()=>{click()}}></button>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/master/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
-  )
+const Index: NextPage = () => {
+return<Home/>
 }
 
 // This gets called on every request
 export async function getServerSideProps() {
-  // Fetch data from external API
-  const res = await fetch(`https://jsonplaceholder.typicode.com/todos/1`)
-  const data = await res.json()
+  
+  const visitorData={
+    id:"visitor_1",
+    context:{
+      age:20
+    },
+  }
 
+  const flagship= Flagship.start(ENV_ID, API_KEY,{
+    fetchNow:false
+  })
+
+  const visitor= flagship?.newVisitor({visitorId: visitorData.id, context: visitorData.context,})
+
+  await visitor?.synchronizeModifications()
 
   // Pass data to the page via props
-  return { props: { campaigns , data, initialModifications: [{
-    key: 'btnColor',
-    campaignId: 'campa',
-    variationGroupId: '',
-    variationId: '',
-    isReference: true,
-    value: 'blue'
-  }]} }
+  return { props: { initialModifications: visitor?.getModificationsArray()||[] } }
 }
-export default Home
+export default Index
